# docs/generated/ranking_openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Ranking Service",
    "version" : "0.0.1"
  },
...
  "components" : {
    "schemas" : {
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "MMRRuleType" : {
        "type" : "string",
        "enum" : [ "FLEX_RANK", "SOLOQ_RANK", "NONE" ]
      },
      "PaginatedResponseDTO" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : { }
          },
          "currentPage" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "QueueDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "name" : {
            "type" : "string"
          },
          "mmrRuleType" : {
            "$ref" : "#/components/schemas/MMRRuleType"
          },
          "initialMMR" : {
            "type" : "integer",
            "format" : "int32"
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
      "Region" : {
        "type" : "string",
        "enum" : [ "BR", "EUNE", "EUW", "LAN", "LAS", "NA", "OCE", "RU", "TR", "ME", "JP", "KR", "SEA", "TW", "VN", "PBE" ]
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
      "UpdateQueueRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : [ "string", "null" ]
          },
          "mmrRuleType" : {
            "$ref" : "#/components/schemas/MMRRuleType"
          },
          "initialMMR" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "active" : {
            "type" : [ "boolean", "null" ]
          }
        }
      }
    }
  },
  "paths" : {
...
    "/api/v1/ranking/leaderboards/queues/{queueId}" : {
      "get" : {
        "summary" : "Get queue leaderboard with pagination",
        "operationId" : "getQueueLeaderboard",
        "tags" : [ "Leaderboard" ],
        "parameters" : [ {
          "description" : "Queue UUID",
          "example" : "550e8400-e29b-41d4-a716-446655440010",
          "name" : "queueId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "description" : "0-based page index",
          "example" : 0,
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "description" : "Optional region filter",
          "name" : "region",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/Region"
          },
          "required" : false
        }, {
          "description" : "Page size (1..100)",
          "example" : 20,
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 0,
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved the leaderboard",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginatedResponseDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid pagination parameters",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Queue not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ranking/queue/{queueId}" : {
      "put" : {
        "summary" : "Update an existing queue's configuration",
        "operationId" : "updateQueue",
        "tags" : [ "Queue Management" ],
        "parameters" : [ {
          "name" : "queueId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateQueueRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Queue configuration successfully updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QueueDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid update payload",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Queue not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get" : {
        "summary" : "Get queue details",
        "operationId" : "getQueue",
        "tags" : [ "Queue Management" ],
        "parameters" : [ {
          "description" : "Queue UUID",
          "example" : "550e8400-e29b-41d4-a716-446655440010",
          "name" : "queueId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Queue found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QueueDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Queue not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete a specific queue",
        "operationId" : "deleteQueue",
        "tags" : [ "Queue Management" ],
        "parameters" : [ {
          "name" : "queueId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Queue successfully deleted"
          },
          "404" : {
            "description" : "Queue not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}